requirements.txt
pandas==2.1.0
requests==2.28.2
beautifulsoup4==4.12.0
pytest==7.3.1
psycopg2-binary==2.9.6
pyspark==3.4.1
